feat: Update version to 2.0.1 and refactor customer group and territory assignment logic
