+++ erpnext_cyprus/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "2.0.0"
+__version__ = "2.0.1"
