+++ erpnext_cyprus/patches.txt
@@ -4,4 +4,3 @@
 
 [post_model_sync]
 # Patches added in this section will be executed after doctypes are migrated
-erpnext_cyprus.patches.update_customer_groups_and_territories
